+++ src/django_core_micha/activity/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ActivityConfig(AppConfig):
+    name = "django_core_micha.activity"
+    label = "django_core_micha_activity"
+    verbose_name = "Platform Activity"